move components to more sane location
