+++ src/components/display/information/Location.tsx
